Trait keystream::SeekableKeyStream

pub trait SeekableKeyStream: KeyStream {
    fn seek_to(&mut self, byte_offset: u64) -> Result<(), Error>;
}

KeyStreams that allow efficiently moving to positions in the stream

Required Methods

fn seek_to(&mut self, byte_offset: u64) -> Result<(), Error>

Seeks to a position, with byte resolution, in the keystream.

Returns an error if the seek would pass the end of the keystream.